Fix bug in FoolCardGame._play_round (game.py)
Bug description:
PLAYERS_NUM=3
If during the round an attacker finishes game and defedner lost it,
then in the next round, defender would attack the 3rd player, which
is wrong.

Solution:
To fix the issue, one was needed to swap places of two methods:
_reassign_roles and _remove_watchers (this is a correct order) in
the end of FoolCardGame._play_round method (game.py).