Replace True for self in Player.defend
Substitution was done in while condition for consistency with other
Player methods (attack and throw_card). It does not really matter if
it's True or self, since number of cards that players can attack with
(or throw to) a defender cannot exceed the number of cards defender
has in the beginning of a round.